refactor: extract CountFilter component to new file
