+++ src/views/components/CountFilter/CountFilter.tsx
@@ -1,0 +1,17 @@
+import { ComponentProps, FormEventHandler } from 'react'
+import style from '@views/components/GamePageList/GamePageList.module.scss'
+
+
+interface ICountFilter extends ComponentProps<'form'> {
+	onHandleSubmit: FormEventHandler
+}
+
+export function CountFilter({ onHandleSubmit }: ICountFilter) {
+	return (
+		<form className={style.CountFilter} onSubmit={onHandleSubmit}>
+			<label htmlFor="item-count">Items per page:</label>
+			<input type="search" name="item-count" id="item-count"/>
+			<button className="primary" type="submit">save</button>
+		</form>
+	)
+}
